fix: refactors form-step container helpers
